Starting with shared mem and sems

+++ src/app.c
@@ -3,16 +3,24 @@
 void create_slaves(int * pipes[][2], pid_t * pids);
 void file_handler(int argc, char * argv[], int * pipes[][2]);
 int check_path(char * path, struct stat fileStat);
-int cycle_pipes(char * argv[], int argc, int * pipes[][2],  struct stat fileStat);
+int cycle_pipes(int argc, char * argv[], int * pipes[][2],  struct stat fileStat);
 void get_results(int * pipes[][2]);
 int send_to_slave(char * argv[], int fd, int * pipes[][2], struct stat fileStat);
 
 static int get_shared_block(char* filename, int size);
-char* attach_memory_block(char* filename, int size);
+void * attach_memory_block(char* filename, int size);
 int detach_memory_block(char* block);
 int destroy_memory_block(char* filename);
+void write_shmem(char * data, int data_size);
 
 static int file_list_iter = 1;
+
+char * shmblock;
+int block_size;
+int used_space = 0;
+
+sem_t * sem_prod;
+sem_t * sem_cons;
 
 int main(int argc, char * argv[]) {
     // "in" and "out" are in reference to the slaves
@@ -40,51 +48,50 @@ int main(int argc, char * argv[]) {
     pid_t slave_pids[SLAVE_COUNT];
     create_slaves(pipes, slave_pids);
 
+    sem_unlink(SEM_PRODUCER_FNAME);
+    sem_unlink(SEM_CONSUMER_FNAME);
+
     // Set up semaphores
-    // sem_t* sem_prod = sem_open(SEM_PRODUCER_FNAME, 0);
-    // if(sem_prod == SEM_FAILED){
-    //     perror("sem_open/producer falied");
-    //     exit(EXIT_FAILURE);
-    // }
+    sem_prod = sem_open(SEM_PRODUCER_FNAME, O_CREAT | O_EXCL, 0644, 0);
+    if(sem_prod == SEM_FAILED){
+        perror("sem_open/producer failed");
+        exit(EXIT_FAILURE);
+    }
 
-    // sem_t* sem_cons = sem_open(SEM_CONSUMER_FNAME, 0);
-    // if(sem_cons == SEM_FAILED){
-    //     perror("sem_open/consumer falied");
-    //     exit(EXIT_FAILURE);
+    sem_cons = sem_open(SEM_CONSUMER_FNAME, O_CREAT | O_EXCL, 0644, 0);
+    if(sem_cons == SEM_FAILED){
+        perror("sem_open/consumer failed");
+        exit(EXIT_FAILURE);
+    }
 
-        file_handler(argc, argv, pipes);
+    //Agregar cuando se lee e imprime: antes de imprimir
+    // Grab the shared memory block
+    block_size = ((argc+1) * (MAX_PATH_LENGTH + MD5_LENGTH + PID_LENGTH + 3));
+    shmblock = attach_memory_block(FILENAME, block_size);
+    if(shmblock == NULL){
+        printf("ERROR: couldn't get block\n");
+        return -1;
+    }
 
 
-    //     //Agregar cuando se lee e imprime: antes de imprimir
-    //     // Grab the shared memory block
-    //     char* block = attach_memory_block(FILENAME, BLOCK_SIZE);
-    //     if(block == NULL){
-    //         printf("ERROR: couldn't get block\n");
-    //         return -1;
-    //     }
-    //     sem_wait(sem_cons); // Wait for the consumer to have an open slot.
-    //     // Escribo y luego:
-    //     sem_post(sem_prod); // Signal tht something has been produced
+    file_handler(argc, argv, pipes);
 
-    //     // Ya fuera del ciclo, luego de terminar
-    //     sem_close(sem_prod);
-    //     sem_close(sem_cons);
-    //     detach_memory_block(block);
-    //     destroy_memory_block(FILENAME);
 
-    //     //Kill the slaves when finished
-    //     for(int i=0; i<SLAVE_COUNT; i++){
-    //         close(pipes[i][0][1]);
-    //         kill(slave_pids[i],SIGTERM);
-    //     }  
-    //     exit(0);
-    // }
+//  sem_wait(sem_prod); // Wait for the consumer to have an open slot.
 
+    // Ya fuera del ciclo, luego de terminar
+    sem_close(sem_prod);
+    sem_close(sem_cons);
+    detach_memory_block(shmblock);
+    destroy_memory_block(FILENAME);
+    //Kill the slaves when finished
     for(int i=0; i<SLAVE_COUNT; i++){
         close(pipes[i][0][1]);
         kill(slave_pids[i],SIGTERM);
     }  
+    exit(0);
     
+
 }
 
 
@@ -116,20 +123,6 @@ void create_slaves(int * pipes[][2], pid_t * pids) {
         else {
             close(pipes[i][0][0]); // Close entry pipe reading fd for app
             close(pipes[i][1][1]); // Close exit pipe writing fd for app
-            // TESTING PIPES:
-            // char* str = "./app";
-            // write(pipes[i][0][1],str,strlen(str)); 
-            // close(pipes[i][0][1]);
-            
-            // waitpid(cpid,NULL,0);
-
-            // char buff[128];
-            // int count;
-            // if((count = read(pipes[i][1][0],buff,sizeof(buff)-1)) > 0){
-            //     buff[count] = '\0';
-            //     printf("%s\n",buff);
-            // }
-            // close(pipes[i][1][0]);
             pids[i] = cpid;
         }
     }
@@ -147,7 +140,7 @@ void file_handler(int argc, char * argv[], int * pipes[][2]) {
         }
     }
     while(file_list_iter < argc) {
-        cycle_pipes(argv, argc, pipes, fileStat);
+        cycle_pipes(argc, argv, pipes, fileStat);
     }
 }
 
@@ -155,6 +148,9 @@ void file_handler(int argc, char * argv[], int * pipes[][2]) {
     Returns -1 on write error or the number written otherwise
 */
 int send_to_slave(char * argv[], int fd, int * pipes[][2], struct stat fileStat) {
+
+  //  sem_wait(sem_prod); // Wait for view.c to consume from buffer
+
     int c, ans = -1;
     if((c = check_path(argv[file_list_iter], fileStat)) == -1){
                 perror("Invalid path");
@@ -194,7 +190,10 @@ int check_path(char * path, struct stat fileStat){
     Cycles through all slave pipes and reads their ouptut, if there is one available. 
     Passes file paths to every slave's pipe.
 */
-int cycle_pipes(char * argv[], int argc, int * pipes[][2], struct stat fileStat) {
+int cycle_pipes(int argc, char * argv[], int * pipes[][2], struct stat fileStat) {
+
+  //  sem_wait(sem_prod); // Wait for view.c to consume from buffer
+
     fd_set writefds;
     int max_fd = 0;
 
@@ -258,60 +257,77 @@ void get_results(int * pipes[][2]){
     for(int i = 0; i < SLAVE_COUNT; i++) {
 
         if(FD_ISSET(pipes[i][1][0], &readfds)) {
-            char buffer[MAX_PATH_LENGTH];
+            char buffer[MAX_PATH_LENGTH + MD5_LENGTH + PID_LENGTH + 3]; // 3 extra spaces for spaces and '\0'
             int count;
-            char md5_hash[MD5_LENGTH]; // Buffer para el hash MD5 (32 caracteres + 1 para '\0')
-            char filename[MAX_PATH_LENGTH]; // Buffer para el nombre del archivo
-            pid_t cpid;
-            if((count = read(pipes[i][1][0], buffer, sizeof(buffer))) == -1) {
+            // char md5_hash[MD5_LENGTH]; // Buffer para el hash MD5 (32 caracteres + 1 para '\0')
+            // char filename[MAX_PATH_LENGTH]; // Buffer para el nombre del archivo
+            if((count = read(pipes[i][1][0], buffer, sizeof(buffer)-1)) == -1) {
                 perror("Read error");
                 exit(EXIT_FAILURE);
             } 
             else {
                buffer[count] = '\0';
-               //Parsear la salida para extraer el hash MD5 y el nombre del archivo
-                if (sscanf(buffer, "%32s %1024s %d", md5_hash, filename, &cpid) == 3) {
+               if(write(STDOUT_FILENO,buffer,count+1) == -1){
+                    perror("Write failed");
+                    exit(1);
+               }
+            //   sem_post(sem_cons); // Signal view.c buffer is ready to be read
 
-                output parsed_data;
-                parsed_data.file_name = filename;
-                parsed_data.md5 = md5_hash;
-                parsed_data.pid = cpid;
+              // write_shmem(buffer,count+1);
+            //    strcat(buffer,cpid);
+            //    printf("%s\n", buffer);
+            //    //Parsear la salida para extraer el hash MD5 y el nombre del archivo
+            //     if (sscanf(buffer, "%32s %1024s %d", md5_hash, filename, &cpid) == 3) {
+
+            //     output parsed_data;
+            //     parsed_data.file_name = filename;
+            //     parsed_data.md5 = md5_hash;
+            //     parsed_data.pid = cpid;
                 
-                printf("path:%s\t\tmd5:%s\tpid:%d\n", parsed_data.file_name , parsed_data.md5, parsed_data.pid);
-                }
-                else {
-                    perror("Error al parsear la salida\n");
-                    exit(EXIT_FAILURE);
-                }  
-            }
+            //     printf("path:%s\t\tmd5:%s\tpid:%d\n", parsed_data.file_name , parsed_data.md5, parsed_data.pid);
+            //     }
+            //     else {
+            //         perror("Error al parsear la salida\n");
+            //         exit(EXIT_FAILURE);
+            //     }  
+             }
         }
     }
-    
 }
 
 // A partir de aca estan las funciones auxiliares de memshare
 
 static int get_shared_block(char* filename, int size){
-    key_t key;
+    FILE *file = fopen(filename, "w");
+    if (file == NULL) {
+        perror("Failed to create temp file");
+        exit(EXIT_FAILURE);
+    }
+    fclose(file);
 
     // Request a key, the key is linked to a filename, so that other programs can access it.
-    key= ftok(filename, 0);
+    key_t key= ftok(filename, 0);
     if(key == IPC_RESULT_ERROR) return IPC_RESULT_ERROR;
+
+    unlink(filename);
 
     //Get shared block --- create it if it does not exist
     return shmget(key, size, 0644 | IPC_CREAT);
 }
 
-char* attach_memory_block(char* filename, int size){
+void * attach_memory_block(char* filename, int size){
     int shared_block_id = get_shared_block(filename, size);
     char* result;
 
-    if(shared_block_id == IPC_RESULT_ERROR) return NULL;
+    if(shared_block_id == IPC_RESULT_ERROR) {
+        return NULL;
+    }
 
-    // Map the shared blovk into this process's memory and give me a pointer to it
+    // Map the shared block into this process's memory and give me a pointer to it
     result = shmat(shared_block_id, NULL, 0);
-    if(result == (char *)IPC_RESULT_ERROR) return NULL;
-
+    if(result == (void *)IPC_RESULT_ERROR){
+        return NULL;
+    }
     return result;
 }
 
@@ -325,4 +341,16 @@ int destroy_memory_block(char * filename){
     if(shared_blok_id == IPC_RESULT_ERROR) return -1;
 
     return (shmctl(shared_blok_id, IPC_RMID, NULL) != IPC_RESULT_ERROR);
+}
+
+void write_shmem(char * data, int data_size){
+    // Verificar si hay suficiente espacio en el bloque
+    if (used_space + data_size > block_size) {
+        printf("No hay suficiente espacio en la memoria compartida.\n");
+        return;
+    }
+
+    // Escribir datos en la memoria compartida
+    memcpy(shmblock + used_space, data, data_size);
+    used_space += data_size;  // Actualizar el espacio utilizado
 }
